% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffAnalysis.R
\name{diffAnalysis}
\alias{diffAnalysis}
\title{Differential peak calling with MeRIP-Seq.}
\usage{
diffAnalysis(
  bam_IP,
  bam_input,
  bam_IP_treated,
  bam_input_treated,
  txdb = NULL,
  genome = NULL,
  bin_size = 25,
  step_size = 25,
  fragment_length = 100,
  paired_end = FALSE,
  strandness = c("unstrand", "1st_strand", "2nd_strand"),
  gff = NULL,
  p_cutoff = 0.001,
  lfc_threshold = 0,
  alt_hypothesis = c("greaterAbs", "lessAbs", "greater", "less"),
  poisson = TRUE,
  save_se = FALSE,
  plot_gc = FALSE,
  parallel = 1
)
}
\value{
the peaks in GRangesList format.
}
\description{
Differential peak calling with MeRIP-Seq.
}
