% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakCalling.R
\name{peakCalling}
\alias{peakCalling}
\title{Peak calling with MeRIP-Seq.}
\usage{
peakCalling(
  bam_IP,
  bam_input,
  txdb = NULL,
  genome = NULL,
  bin_size = 25,
  step_size = 25,
  fragment_length = 100,
  paired_end = FALSE,
  strandness = c("unstrand", "1st_strand", "2nd_strand"),
  gff = NULL,
  p_cutoff = 1e-10,
  poisson = TRUE,
  plot_gc = FALSE,
  parallel = 1
)
}
\value{
the peaks in GRangesList format.
}
\description{
Peak calling with MeRIP-Seq.
}
