% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotGCbias}
\alias{plotGCbias}
\title{Plot Poisson GLM fits for GC content bias detection}
\usage{
plotGCbias(se, savePrefix = "gc_bias")
}
\value{
save PDF files of GC bias curves
}
\description{
Plot Poisson GLM fits for GC content bias detection
}
