% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTopology}
\alias{plotTopology}
\title{Plot transcript topology distribution for peaks}
\usage{
plotTopology(list_grl, txdb, savePrefix = "topology")
}
\value{
save PDF files of tx topology
}
\description{
Plot transcript topology distribution for peaks
}
